% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Data Set}
\format{
## `fire`
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{zDistance fire traveled}
  \item{DAMAGE}{damage}
}
}
\source{
Textbook
}
\usage{
fire
}
\description{
A data set containing the distance and damage from different fires.
}
\details{
15 lines of measurements
}
\keyword{datasets}
