% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Experiment Simulator}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of binomial experiements to iterate}

\item{n}{Number of Bernoulli experiments per binomial}

\item{p}{Probability of success}
}
\value{
Produces a bar graph that depicts the frequency of the number of successes
}
\description{
Binomial Experiment Simulator
}
\examples{
\dontrun{mybin()}
}
