% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap interval estimator}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of times the experiment is to be repeated}

\item{x}{The data}

\item{fun}{Function of the sample to be estimated}

\item{alpha}{1-alpha = ci}

\item{cx}{The cx value}

\item{...}{}
}
\value{
A graph, a list of output parameters
}
\description{
Bootstrap interval estimator
}
\examples{
\dontrun{myboot2(iter = 10000, x, fun = "mean", alpha = 0.05)}
}
