% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Mycltu from lab *}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{size of random sample}

\item{iter}{number of random samples}

\item{a}{lower bound of random sample}

\item{b}{upper bound of random sample}
}
\value{
A sick graph
}
\description{
Mycltu from lab *
}
\examples{
\dontrun{mycltu(10,10000,0,5)}
}
