% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Curve Generator and Lower Tail calculator}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution}

\item{sigma}{Standard deviation of the normal distribution}

\item{a}{quantile}
}
\value{
Produces a curve with the lower tail shaded and returns the calculated lower tail probability
}
\description{
Normal Curve Generator and Lower Tail calculator
}
\examples{
\dontrun{myncurve(0,2,1)}
}
