% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myntickets.R
\name{myntickets}
\alias{myntickets}
\title{Airplane ticket problem function}
\usage{
myntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{Number of seats, default 200}

\item{gamma}{Probability each passenger is a no-show, default 0.02}

\item{p}{Desired probability that N or fewer passengers show up, default 0.95}
}
\value{
Produces graphs for the objective function vs n for both the binomial calculation and the normal distribution. Also prints a named list of nd, nc, N, p, and gamma.
}
\description{
Airplane ticket problem function
}
\examples{
\dontrun{myntickets()}
}
